feat: Add d455 as stereo cam to sensors xacro (#20)
Together with the branch of the same name in the sensors-camera repo,
this modifies the standard launch to also launch stereo images for the
d455. It is currently not possible to selectively remove this sensor
using flags at launch because the chain of parameter passing is broken
at some link.